moved some cred repo classes to a sub-namespace
