+++ mRemoteNGTests/Credential/CredentialRepositoryListTests.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using mRemoteNG.Credential;
+using mRemoteNG.Credential.Repositories;
 using NSubstitute;
 using NUnit.Framework;
