+++ mRemoteV1/Credential/Repositories/CompositeRepositoryUnlocker.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security;
 
-namespace mRemoteNG.Credential
+namespace mRemoteNG.Credential.Repositories
 {
     public class CompositeRepositoryUnlocker
     {
