+++ mRemoteV1/mRemoteV1.csproj
@@ -215,11 +215,11 @@
     <Compile Include="Connection\Protocol\Http\Connection.Protocol.HTTPS.CertEvent.cs" />
     <Compile Include="Connection\Protocol\ProtocolFactory.cs" />
     <Compile Include="Connection\Protocol\VNC\VNCEnum.cs" />
-    <Compile Include="Credential\CompositeRepositoryUnlocker.cs" />
+    <Compile Include="Credential\Repositories\CompositeRepositoryUnlocker.cs" />
     <Compile Include="Credential\CredentialChangedEventArgs.cs" />
     <Compile Include="Credential\CredentialDeletionMsgBoxConfirmer.cs" />
     <Compile Include="Credential\CredentialDomainUserComparer.cs" />
-    <Compile Include="Credential\CredentialRepoUnlockerBuilder.cs" />
+    <Compile Include="Credential\Repositories\CredentialRepoUnlockerBuilder.cs" />
     <Compile Include="Credential\CredentialServiceFactory.cs" />
     <Compile Include="Credential\CredentialsService.cs" />
     <Compile Include="Security\Factories\CryptoProviderFactoryFromSettings.cs" />
@@ -230,11 +230,11 @@
     <Compile Include="Security\RandomGenerator.cs" />
     <Compile Include="Tools\CustomCollections\IFullyNotifiableList.cs" />
     <Compile Include="Tools\CustomCollections\FullyObservableCollection.cs" />
-    <Compile Include="Credential\CredentialRepositoryChangedArgs.cs" />
-    <Compile Include="Credential\CredentialRepositoryList.cs" />
+    <Compile Include="Credential\Repositories\CredentialRepositoryChangedArgs.cs" />
+    <Compile Include="Credential\Repositories\CredentialRepositoryList.cs" />
     <Compile Include="Credential\CredentialRecord.cs" />
     <Compile Include="Credential\CredentialRecordTypeConverter.cs" />
-    <Compile Include="Credential\CredentialRepositoryConfig.cs" />
+    <Compile Include="Credential\Repositories\CredentialRepositoryConfig.cs" />
     <Compile Include="Credential\ICredentialRecord.cs" />
     <Compile Include="Credential\ICredentialRepositoryList.cs" />
     <Compile Include="Credential\ICredentialRepository.cs" />
